Fix empty file inputs in Symfony 4.1
Symfony 4.1 fails on uploaded files that are null: they must be array
or UploadedFile. UploadedFile can only be used for actual files.
Solution is to pass the uploaded file as array instead.

The array cannot be passed for actual uploaded files because Symfony
will not trust the files because they were not created by PHP core.

A better solution would be to let Symfony accept null files (see
implementation of FileBag).